+++ packages/Stun/lib/stun.dart
@@ -6,4 +6,5 @@ library stun;
 export 'src/types/stun_types.dart';
 export 'src/interfaces/i_stun_handler.dart';
 export 'src/implementations/stun_handler.dart';
+export 'src/implementations/stun_handler_singleton.dart';
 export 'src/implementations/nat_detector.dart';